improved the functionality of the buff mode and also fixed a bug in writeflash()

+++ user/customized_type.h
@@ -3,7 +3,7 @@
 #define CUSTOMIZED_TYPE_H
 typedef union u32ANDint16_t{
 	u32 flash;
-	int16_t mem;
+	float mem;
 } u32ANDint16_t;
 typedef enum CarMode
 {
